requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML and embeddings
openai==1.3.0
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0

# Vector database
chromadb==0.4.15

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0

# Text processing and NLP
spacy==3.7.2
nltk==3.8.1

# Data handling
pandas==2.1.3
numpy==1.25.2

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Validation
pydantic==2.5.0

# Logging
loguru==0.7.2